requirements.txt
Adafruit-Blinka==8.38.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-neopixel==6.3.11
adafruit-circuitpython-pixelbuf==2.0.4
adafruit-circuitpython-requests==3.2.2
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
numpy==1.26.4
PyAudio==0.2.14
pyftdi==0.55.0
pyqtgraph==0.13.6
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
scipy==1.12.0
sysv-ipc==1.1.0
typing-extensions==4.10.0
